Replace old read-only checker with new implementation.
gcc/rust/ChangeLog:

	* Make-lang.in: Replace old read-only checker with new implementation.
	* checks/errors/rust-readonly-check.cc (ReadonlyChecker::ReadonlyChecker):
	Replace old read-only checker with new implementation.
	* checks/errors/rust-readonly-check.h:
	Replace old read-only checker with new implementation.
	* rust-session-manager.cc (Session::compile_crate):
	Switch to new read-only checker.
	* checks/errors/rust-readonly-check2.cc: Removed.
	* checks/errors/rust-readonly-check2.h: Removed.
